Refactoring mostly complete (lacking more docstrings)
